Dow Jones Industrial Average (DJIA)

Dow Jones Industrial Average (DJIA) is a stock market index that shows how 30 large publicly owned companies based in the United States have traded during a standard NYSE trading session. We consider monthly cross-sectional returns from April 2004 to December 2017. The data were obtained from the CRSP (Center for Research in Security Prices) database.

data("DJI_return")

Format

A data matrix

References

Kokoszka, P., Miao, H., Petersen, A. and Shang, H. L. (2019) `Forecasting of density functions with an application to cross-sectional and intraday returns', International Journal of Forecasting, 35(4), 1304-1317.

Examples

data(DJI_return)