Dynamic functional principal component analysis for density forecasting

Implementation of a dynamic functional principal component analysis to forecast densities.

Horta_Ziegelmann_FPCA(data, gridpoints, h_scale = 1, p = 5, m = 5001, 
  kernel = c("gaussian", "epanechnikov"), band_choice = c("Silverman", "DPI"), 
  VAR_type = "both", lag_maximum = 6, no_boot = 1000, alpha_val = 0.1, 
  ncomp_select = "TRUE", D_val = 10)

Arguments

data

Densities or raw data matrix of dimension N by p, where N denotes sample size and p denotes dimensionality

gridpoints

Grid points

h_scale

Scaling parameter in the kernel density estimator

p

Number of backward parameters

m

Number of grid points

kernel

Type of kernel function

band_choice

Selection of optimal bandwidth

VAR_type

Type of vector autoregressive process

lag_maximum

A tuning parameter in the super_fun function

no_boot

A tuning parameter in the super_fun function

alpha_val

A tuning parameter in the super_fun function

ncomp_select

A tuning parameter in the super_fun function

D_val

A tuning parameter in the super_fun function

Details

1) Compute a kernel covariance function 2) Via eigen-decomposition, a density can be decomposed into a set of functional principal components and their associated scores 3) Fit a vector autoregressive model to the scores with the order selected by Akaike information criterion 4) By multiplying the estimated functional principal components with the forecast scores, obtain forecast densities 5) Since forecast densities may neither be non-negative nor sum to one, normalize the forecast densities accordingly

Value

Yhat.fix_den

Forecast density

u

Grid points

du

Distance between two successive grid points

Ybar_est

Mean of density functions

psihat_est

Estimated functional principal components

etahat_est

Estimated principal component scores

etahat_pred_val

Forecast principal component scores

selected_d0

Selected number of components

selected_d0_pvalues

p-values associated with the selected functional principal components

thetahat_val

Estimated eigenvalues

References

Horta, E. and Ziegelmann, F. (2018) `Dynamics of financial returns densities: A functional approach applied to the Bovespa intraday index', International Journal of Forecasting, 34, 75-88.

Bathia, N., Yao, Q. and Ziegelmann, F. (2010) `Identifying the finite dimensionality of curve time series', The Annals of Statistics, 38, 3353-3386.

Author

Han Lin Shang

See also

Examples

if (FALSE) { # \dontrun{
Horta_Ziegelmann_FPCA(data = DJI_return, kernel = "epanechnikov", 
        band_choice = "DPI", ncomp_select = "FALSE")
} # }