Multilevel functional principal component analysis for clustering

A multilevel functional principal component analysis for performing clustering analysis

MFPCA(y, M = NULL, J = NULL, N = NULL)

Arguments

y

A data matrix containing functional responses. Each row contains measurements from a function at a set of grid points, and each column contains measurements of all functions at a particular grid point

M

Number of countries

J

Number of functional responses in each country

N

Number of grid points per function

Value

K1

Number of components at level 1

K2

Number of components at level 2

K3

Number of components at level 3

lambda1

A vector containing all level 1 eigenvalues in non-increasing order

lambda2

A vector containing all level 2 eigenvalues in non-increasing order

lambda3

A vector containing all level 3 eigenvalues in non-increasing order

phi1

A matrix containing all level 1 eigenfunctions. Each row contains an eigenfunction evaluated at the same set of grid points as the input data. The eigenfunctions are in the same order as the corresponding eigenvalues

phi2

A matrix containing all level 2 eigenfunctions. Each row contains an eigenfunction evaluated at the same set of grid points as the input data. The eigenfunctions are in the same order as the corresponding eigenvalues

phi3

A matrix containing all level 3 eigenfunctions. Each row contains an eigenfunction evaluated at the same set of grid points as the input data. The eigenfunctions are in the same order as the corresponding eigenvalues

scores1

A matrix containing estimated level 1 principal component scores. Each row corresponds to the level 1 scores for a particular subject in a cluster. The number of rows is the same as that of the input matrix y. Each column contains the scores for a level 1 component for all subjects

scores2

A matrix containing estimated level 2 principal component scores. Each row corresponds to the level 2 scores for a particular subject in a cluster. The number of rows is the same as that of the input matrix y. Each column contains the scores for a level 2 component for all subjects.

scores3

A matrix containing estimated level 3 principal component scores. Each row corresponds to the level 3 scores for a particular subject in a cluster. The number of rows is the same as that of the input matrix y. Each column contains the scores for a level 3 component for all subjects.

mu

A vector containing the overall mean function

eta

A matrix containing the deviation from overall mean function to country-specific mean function. The number of rows is the number of countries

Rj

Common trend

Uij

Country-specific mean function

Author

Chen Tang, Yanrong Yang and Han Lin Shang

See also