Testing stationarity of functional time series

A hypothesis test for stationarity of functional time series.

T_stationary(sample, L = 49, J = 500, MC_rep = 1000, cumulative_var = .90,
        Ker1 = FALSE, Ker2 = TRUE, h = ncol(sample)^.5, pivotal = FALSE,
        use_table = FALSE, significance)

Arguments

sample

A matrix of discretised curves of dimension (p by n), where p represents the dimensionality and n represents sample size.

L

Number of Fourier basis functions.

J

Truncation level used to approximate the distribution of the squared integrals of Brownian bridges that appear in the limit distribution.

MC_rep

Number of replications.

cumulative_var

Amount of variance explained.

Ker1

Flat top kernel in (4.1) of Horvath et al. (2014).

Ker2

Flat top kernel in (7) of Politis (2003).

h

Kernel bandwidth.

pivotal

If pivotal = TRUE, a pivotal statistic is used.

use_table

If use_table = TRUE, use the critical values that are available in the book titled Inference for Functional Data (Table 6.1, page 88).

significance

Level of significance. Possibilities are “10%”, “5%”, “1%”.

Details

As in traditional (scalar and vector) time series analysis, many inferential procedures for functional time series assume stationarity. Stationarity is required for functional dynamic regression models, for bootstrap and resampling methods for functional time series and for the functional analysis of volatility.

Value

p-value

When p-value is less than any level of significance, we reject the null hypothesis and conclude that the tested functional time series is not stationary.

References

L. Horvath and Kokoszka, P. (2012) Inference for Functional Data with Applications, Springer, New York.

L. Horvath, P. Kokoszka, G. Rice (2014) "Testing stationarity of functional time series", Journal of Econometrics, 179(1), 66-82.

D. N. Politis (2003) "Adaptive bandwidth choice", Journal of Nonparametric Statistics, 15(4-5), 517-533.

A. Aue, G. Rice, O. S\"onmez (2018) "Detecting and dating structural breaks in functional data without dimension reduction", Journal of the Royal Statistical Society: Series B, 80(3), 509-529.

Author

Greg. Rice and Han Lin Shang

See also

Examples

result = T_stationary(sample = pm_10_GR_sqrt$y)
#> 
#> Monte Carlo test of stationarity of a functional time series
#> 
#> null hypothesis: the series is stationary
#> 
#> p-value = 0.077 
#> N (number of functions) = 182 
#> number of MC replications = 1000
result_pivotal = T_stationary(sample = pm_10_GR_sqrt$y, J = 100, MC_rep = 5000,
        h = 20, pivotal = TRUE)
#> 
#> Pivotal test of stationarity for a functional time series
#> 
#> null hypothesis: the series is stationary
#> 
#> p-value = 0.1208 
#> N (number of functions) = 182 
#> number of MC replications = 5000