Forecasting via a high-dimensional functional principal component regression

Forecast high-dimensional functional principal component model.

# S3 method for class 'hdfpca'
forecast(object, h = 3, level = 80, B = 50, ...)

Arguments

object

An object of class 'hdfpca'

h

Forecast horizon

level

Prediction interval level, the default is 80 percent

B

Number of bootstrap replications

...

Other arguments passed to forecast routine.

Details

The low-dimensional factors are forecasted with autoregressive integrated moving average (ARIMA) models separately. The forecast functions are then calculated using the forecast factors. Bootstrap prediction intervals are constructed by resampling from the forecast residuals of the ARIMA models.

Value

forecast

A list containing the h-step-ahead forecast functions for each population

upper

Upper confidence bound for each population

lower

Lower confidence bound for each population

References

Y. Gao, H. L. Shang and Y. Yang (2018) High-dimensional functional time series forecasting: An application to age-specific mortality rates, Journal of Multivariate Analysis, forthcoming.

Author

Y. Gao and H. L. Shang

See also

Examples

if (FALSE) { # \dontrun{
hd_model = hdfpca(hd_data, order = 2, r = 2)
hd_model_fore = forecast.hdfpca(object = hd_model, h = 1)
} # }