Forecast functional time series

The coefficients from the fitted object are forecasted using either an ARIMA model (method = "arima"), an AR model (method = "ar"), an exponential smoothing method (method = "ets"), a linear exponential smoothing method allowing missing values (method = "ets.na"), or a random walk with drift model (method = "rwdrift"). The forecast coefficients are then multiplied by the principal components to obtain a forecast curve.

ftsmiterativeforecasts(object, components, iteration = 20)

Arguments

object

An object of class fts.

components

Number of principal components.

iteration

Number of iterative one-step-ahead forecasts.

Details

1. Obtain a smooth curve \(f_t(x)\) for each \(t\) using a nonparametric smoothing technique.

2. Decompose the smooth curves via a functional principal component analysis.

3. Fit a univariate time series model to each of the principal component scores.

4. Forecast the principal component scores using the fitted time series models.

5. Multiply the forecast principal component scores by fixed principal components to obtain forecasts of \(f_{n+h}(x)\).

6. The estimated variances of the error terms (smoothing error and model residual error) are used to compute prediction intervals for the forecasts.

Value

List with the following components:

mean

An object of class fts containing point forecasts.

lower

An object of class fts containing lower bound for prediction intervals.

upper

An object of class fts containing upper bound for prediction intervals.

fitted

An object of class fts of one-step-ahead forecasts for historical data.

error

An object of class fts of one-step-ahead errors for historical data.

coeff

List of objects of type forecast containing the coefficients and their forecasts.

coeff.error

One-step-ahead forecast errors for each of the coefficients.

var

List containing the various components of variance: model, error, mean, total and coeff.

model

Fitted ftsm model.

bootsamp

An array of \(dim = c(p, B, h)\) containing the bootstrapped point forecasts. \(p\) is the number of variables. \(B\) is the number of bootstrap samples. \(h\) is the forecast horizon.

References

H. Booth and R. J. Hyndman and L. Tickle and P. D. Jong (2006) "Lee-Carter mortality forecasting: A multi-country comparison of variants and extensions", Demographic Research, 15, 289-310.

B. Erbas and R. J. Hyndman and D. M. Gertig (2007) "Forecasting age-specific breast cancer mortality using functional data model", Statistics in Medicine, 26(2), 458-470.

R. J. Hyndman and M. S. Ullah (2007) "Robust forecasting of mortality and fertility rates: A functional data approach", Computational Statistics and Data Analysis, 51(10), 4942-4956.

R. J. Hyndman and H. Booth (2008) "Stochastic population forecasts using functional data models for mortality, fertility and migration", International Journal of Forecasting, 24(3), 323-342.

R. J. Hyndman and H. L. Shang (2009) "Forecasting functional time series" (with discussion), Journal of the Korean Statistical Society, 38(3), 199-221.

Author

Han Lin Shang

See also

Examples

# Iterative one-step-ahead forecasts via functional principal component analysis.  
ftsmiterativeforecasts(object = Australiasmoothfertility, components = 2, iteration = 5)
#> Functional time series
#>  y: Smoothed fertility rate
#>  x: Age