Simulated high-dimensional functional time series

We generate \(N\) populations of functional time series. For each \(i\in \{1,\dots, N\}\), the \(i\)th function at time \(t\in \{1,\dots, T\}\) is given by $$X_t^{(i)}(u) = \sum^2_{p=1}\beta_{p,t}^{(i)}\gamma_p^{(i)}(u) + \theta_t^{(i)}(u),$$ where \(\theta_t^{(i)}(u) = \sum^{\infty}_{p=3}\beta_{p,t}^{(i)}\gamma_p^{(i)}(u)\).

data("hd_data")

Details

The coefficients \(\beta_{p,t}^{(i)}\) for all \(N\) populations are combined and generated, for all \(p\in N\), by $$\bm{\beta}_{p,t} = \bm{A}_p\bm{f}_{p,t},$$ where \(\bm{\beta}_{p,t}=\{\beta_{p,t}^{1},\dots,\beta_{p,t}^N\}\). Here, \(\bm{A}_p\) is an \(N\times N\) matrix, and \(\bm{f}_{p,t}\) is an \(N\times 1\) vector. Furthermore, we assume that the \(\beta_{p,t}^{(i)}\)s have mean 0 and variance 0 when \(p>3\), so we only construct the coefficients \(\bm{\beta}_{p,t}\) for \(p\in\{1, 2, 3\}\).

The first set of coefficients \(\bm{\beta}_{1,t}\) for \(N\) populations are generated with \(\bm{\beta}_{1,t}=\bm{A}_1\bm{f}_{1,t}\). Each element in the matrix \(\bm{A}_1\) is generated by \(a_{ij}=N^{-1/4}\times b_{ij}\), where \(b_{ij}\sim N(2,4)\).

The factors \(\bm{f}_{1,t}\) are generated using an autoregressive model of order 1, i.e., AR(1). Define the \(i\)th element in vector \(\bm{f}_{1,t}\) as \(f_{1,t}^{(i)}\). Then, \(f_{1,t}^{1}\) is generated by \(f_{1,t}^{1}=0.5\times f_{1,t-1}^{1}+\omega_t\), where \(\omega_t\) are independent \(N(0,1)\) random variables. We generate \(f_{1,t}^{(i)}\) for all \(i\in \{2,\dots, N\}\) by \(f_{1,t}^{(i)}=(1/N) \times g_t^{(i)}\), where \(g_t^{(2)},\dots,g_t^{(N)}\) are also AR(1) and follow \(g_t^{(i)} = 0.2\times g_{t-1}^{(i)}+\omega_t\). It is then ensured that most of the variance of \(\bm{\beta}_{1,t}\) can be explained by one factor. The second coefficient \(\bm{\beta}_{2,t}\) are constructed the same way as \(\bm{\beta}_{1,t}\).

We also generate the third functional principal component scores \(\bm{\beta}_{3,t}\) but with small values. Moreover, \(\bm{A}_3\) is generated by \(a_{ij}=N^{-1/4}\times b_{ij}\), where \(b_{ij}\sim N(0, 0.04)\). The factors \(bm{f}_{3,t}\) are generated as \(\bm{f}_{1,t}\).

The three basis functions are constructed by \(\gamma_1^{(i)}(u) = \sin(2\pi u + \pi i/2)\), \(\gamma_2^{(i)}(u) = \cos(2\pi u + \pi i/2)\) and \(\gamma_3^{(i)}(u) = \sin(4\pi u + \pi i/2)\), where \(u\in [0,1]\). Finally, the functional time series for the \(i\)th population is constructed by $$\bm{X}_t^{(i)}(u) = \bm{\beta}_{1,t}\gamma_1^{(i)}(u) + \bm{\beta}_{2,t}\gamma_2^{(i)}(u) + \bm{\beta}_{3,t}\gamma_3^{(i)}(u),$$ where \((\cdot)_i\) denotes the \(i\)th element of the vector.

References

Y. Gao, H. L. Shang and Y. Yang (2018) High-dimensional functional time series forecasting: An application to age-specific mortality rates, Journal of Multivariate Analysis, forthcoming.

See also

Examples

data(hd_data)