High-dimensional functional principal component analysis

Fit a high dimensional functional principal component analysis model to a multiple-population of functional time series data.

hdfpca(y, order, q = sqrt(dim(y[[1]])[2]), r)

Arguments

y

A list, where each item is a population of functional time series. Each item is a data matrix of dimension p by n, where p is the number of discrete points in each function and n is the sample size

order

The number of principal component scores to retain in the first step dimension reduction

q

The tuning parameter used in the first step dimension reduction, by default it is equal to the square root of the sample size

r

The number of factors to retain in the second step dimension reduction

Details

In the first step, dynamic functional principal component analysis is performed on each population and then in the second step, factor models are fitted to the resulting principal component scores. The high-dimensional functional time series are thus reduced to low-dimensional factors.

Value

y

The input data

p

The number of discrete points in each function

fitted

A list containing the fitted functions for each population

m

The number of populations

model

Model 1 includes the first step dynamic functional principal component analysis models, model 2 includes the second step high-dimensional principal component analysis models

order

Input order

r

Input r

References

Y. Gao, H. L. Shang and Y. Yang (2018) High-dimensional functional time series forecasting: An application to age-specific mortality rates, Journal of Multivariate Analysis, forthcoming.

Author

Y. Gao and H. L. Shang

See also

Examples

hd_model = hdfpca(hd_data, order = 2, r = 2)