Integrated Squared Forecast Error for models of various orders

Computes integrated squared forecast error (ISFE) values for functional time series models of various orders.

isfe.fts(data, max.order = N - 3, N = 10, h = 5:10, method = 
 c("classical", "M", "rapca"), mean = TRUE, level = FALSE, 
  fmethod = c("arima", "ar", "ets", "ets.na", "struct", "rwdrift", 
   "rw", "arfima"), lambda = 3, ...)

Arguments

data

An object of class fts.

max.order

Maximum number of principal components to fit.

N

Minimum number of functional observations to be used in fitting a model.

h

Forecast horizons over which to average.

method

Method to use for principal components decomposition. Possibilities are “M”, “rapca” and “classical”.

mean

Indicates if mean term should be included.

level

Indicates if level term should be included.

fmethod

Method used for forecasting. Current possibilities are “ets”, “arima”, “ets.na”, “struct”, “rwdrift” and “rw”.

lambda

Tuning parameter for robustness when method = "M".

...

Additional arguments controlling the fitting procedure.

Value

Numeric matrix with (max.order+1) rows and length(h) columns containing ISFE values for models of orders 0:(max.order).

Author

Rob J Hyndman

References

R. J. Hyndman and M. S. Ullah (2007) "Robust forecasting of mortality and fertility rates: A functional data approach", Computational Statistics and Data Analysis, 51(10), 4942-4956.

Note

This function can be very time consuming for data with large dimensionality or large sample size. By setting max.order small, computational speed can be dramatically increased.

See also