Estimating long-run covariance function for a functional time series

Bandwidth estimation in the long-run covariance function for a functional time series, using different types of kernel function

long_run_covariance_estimation(dat, C0 = 3, H = 3)

Arguments

dat

A matrix of p by n, where p denotes the number of grid points and n denotes sample size

C0

A tuning parameter used in the adaptive bandwidth selection algorithm of Rice

H

A tuning parameter used in the adaptive bandwidth selection algorithm of Rice

Value

An estimated covariance function of size (p by p)

References

L. Horvath, G. Rice and S. Whipple (2016) Adaptive bandwidth selection in the long run covariance estimation of functional time series, Computational Statistics and Data Analysis, 100, 676-693.

G. Rice and H. L. Shang (2017) A plug-in bandwidth selection procedure for long run covariance estimation with stationary functional time series, Journal of Time Series Analysis, 38(4), 591-609.

D. Li, P. M. Robinson and H. L. Shang (2018) Long-range dependent curve time series, Journal of the American Statistical Association: Theory and Methods, under revision.

Author

Han Lin Shang

See also

Examples

dum = long_run_covariance_estimation(dat = ElNino_OISST_region_1and2$y[,1:5])