Median functions for functional time series

Computes median of functional time series at each variable.

# S3 method for class 'fts'
median(x, na.rm, method = c("hossjercroux", "coordinate", "FM", "mode", 
 "RP", "RPD", "radius"), alpha, beta, weight, ...)

Arguments

x

An object of class fts.

na.rm

Remove any missing value.

method

Method for computing median.

alpha

Tuning parameter when method="radius".

beta

Trimming percentage, by default it is 0.25, when method="radius".

weight

Hard thresholding or soft thresholding.

...

Other arguments.

Details

If method = "coordinate", it computes a coordinate-wise median.

If method = "hossjercroux", it computes the L1-median using the Hossjer-Croux algorithm.

If method = "FM", it computes the median of trimmed functional data ordered by the functional depth of Fraiman and Muniz (2001).

If method = "mode", it computes the median of trimmed functional data ordered by \(h\)-modal functional depth.

If method = "RP", it computes the median of trimmed functional data ordered by random projection depth.

If method = "RPD", it computes the median of trimmed functional data ordered by random projection derivative depth.

If method = "radius", it computes the mean of trimmed functional data ordered by the notion of alpha-radius.

Value

A list containing x = variables and y = median rates.

References

O. Hossjer and C. Croux (1995) "Generalized univariate signed rank statistics for testing and estimating a multivariate location parameter", Journal of Nonparametric Statistics, 4(3), 293-308.

A. Cuevas and M. Febrero and R. Fraiman (2006) "On the use of bootstrap for estimating functions with functional data", Computational Statistics and Data Analysis, 51(2), 1063-1074.

A. Cuevas and M. Febrero and R. Fraiman (2007), "Robust estimation and classification for functional data via projection-based depth notions", Computational Statistics, 22(3), 481-496.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2007) "A functional analysis of NOx levels: location and scale estimation and outlier detection", Computational Statistics, 22(3), 411-427.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2008) "Outlier detection in functional data by depth measures, with application to identify abnormal NOx levels", Environmetrics, 19(4), 331-345.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2010) "Measures of influence for the functional linear model with scalar response", Journal of Multivariate Analysis, 101(2), 327-339.

J. A. Cuesta-Albertos and A. Nieto-Reyes (2010) "Functional classification and the random Tukey depth. Practical issues", Combining Soft Computing and Statistical Methods in Data Analysis, Advances in Intelligent and Soft Computing, 77, 123-130.

D. Gervini (2012) "Outlier detection and trimmed estimation in general functional spaces", Statistica Sinica, 22(4), 1639-1660.

Author

Rob J Hyndman, Han Lin Shang

See also

Examples

# Calculate the median function by the different depth measures.  
median(x = ElNino_ERSST_region_1and2, method = "hossjercroux")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>        1        2        3        4        5        6        7        8 
#> 24.20865 25.70742 26.07100 25.11941 23.83259 22.56601 21.46505 20.56117 
#>        9       10       11       12 
#> 20.30941 20.59353 21.28436 22.44395 
#> 
median(x = ElNino_ERSST_region_1and2, method = "coordinate")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>     1     2     3     4     5     6     7     8     9    10    11    12 
#> 24.19 25.70 26.02 25.12 23.85 22.40 21.46 20.60 20.43 20.54 21.37 22.41 
#> 
median(x = ElNino_ERSST_region_1and2, method = "FM")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>     1     2     3     4     5     6     7     8     9    10    11    12 
#> 24.32 25.68 26.48 25.54 23.98 22.76 21.12 20.52 20.43 20.52 21.24 22.41 
#> 
median(x = ElNino_ERSST_region_1and2, method = "mode")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>     1     2     3     4     5     6     7     8     9    10    11    12 
#> 24.09 25.94 26.04 25.02 24.18 22.83 21.30 20.57 20.24 20.31 21.00 22.31 
#> 
median(x = ElNino_ERSST_region_1and2, method = "RP")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>     1     2     3     4     5     6     7     8     9    10    11    12 
#> 24.09 25.94 26.04 25.02 24.18 22.83 21.30 20.57 20.24 20.31 21.00 22.31 
#> 
median(x = ElNino_ERSST_region_1and2, method = "RPD")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>     1     2     3     4     5     6     7     8     9    10    11    12 
#> 24.80 25.72 26.02 24.95 23.49 22.52 21.46 20.10 19.61 20.54 21.19 22.26 
#> 
median(x = ElNino_ERSST_region_1and2, method = "radius", 
  alpha = 0.5, beta = 0.25, weight = "hard")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>     1     2     3     4     5     6     7     8     9    10    11    12 
#> 24.15 25.66 25.84 24.81 23.48 22.19 21.17 20.26 20.07 20.37 21.09 22.26 
#> 
median(x = ElNino_ERSST_region_1and2, method = "radius", 
  alpha = 0.5, beta = 0.25, weight = "soft")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>      1      2      3      4      5      6      7      8      9     10     11 
#> 24.165 25.645 25.860 24.815 23.485 22.185 21.240 20.350 20.030 20.340 21.075 
#>     12 
#> 22.310 
#>