Standard deviation functions for functional time series

Computes standard deviation of functional time series at each variable.

# S3 method for class 'fts'
sd(x, method = c("coordinate", "FM", "mode", "RP", "RPD", "radius"), 
 trim = 0.25, alpha, weight,...)

Arguments

x

An object of class fts.

method

Method for computing median.

trim

Percentage of trimming.

alpha

Tuning parameter when method="radius".

weight

Hard thresholding or soft thresholding.

...

Other arguments.

Details

If method = "coordinate", it computes coordinate-wise standard deviation functions.

If method = "FM", it computes the standard deviation functions of trimmed functional data ordered by the functional depth of Fraiman and Muniz (2001).

If method = "mode", it computes the standard deviation functions of trimmed functional data ordered by \(h\)-modal functional depth.

If method = "RP", it computes the standard deviation functions of trimmed functional data ordered by random projection depth.

If method = "RPD", it computes the standard deviation functions of trimmed functional data ordered by random projection with derivative depth.

If method = "radius", it computes the standard deviation function of trimmed functional data ordered by the notion of alpha-radius.

Value

A list containing x = variables and y = standard deviation rates.

References

O. Hossjer and C. Croux (1995) "Generalized univariate signed rank statistics for testing and estimating a multivariate location parameter", Nonparametric Statistics, 4(3), 293-308.

A. Cuevas and M. Febrero and R. Fraiman (2006) "On the use of bootstrap for estimating functions with functional data", Computational Statistics and Data Analysis, 51(2), 1063-1074.

A. Cuevas and M. Febrero and R. Fraiman (2007), "Robust estimation and classification for functional data via projection-based depth notions", Computational Statistics, 22(3), 481-496.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2007) "A functional analysis of NOx levels: location and scale estimation and outlier detection", Computational Statistics, 22(3), 411-427.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2008) "Outlier detection in functional data by depth measures, with application to identify abnormal NOx levels", Environmetrics, 19(4), 331-345.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2010) "Measures of influence for the functional linear model with scalar response", Journal of Multivariate Analysis, 101(2), 327-339.

J. A. Cuesta-Albertos and A. Nieto-Reyes (2010) "Functional classification and the random Tukey depth. Practical issues", Combining Soft Computing and Statistical Methods in Data Analysis, Advances in Intelligent and Soft Computing, 77, 123-130.

D. Gervini (2012) "Outlier detection and trimmed estimation in general functional spaces", Statistica Sinica, 22(4), 1639-1660.

Author

Han Lin Shang

See also

Examples

# Fraiman-Muniz depth was arguably the oldest functional depth.  
sd(x = ElNino_ERSST_region_1and2, method = "FM")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.5312960 0.5370651 0.7111353 0.9363543 0.9587743 0.9165644 0.7982182 0.7575417 
#>         9        10        11        12 
#> 0.6939671 0.7263353 0.8210953 0.8217938 
#> 
sd(x = ElNino_ERSST_region_1and2, method = "coordinate")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.9051154 0.7756023 0.9115281 1.1485897 1.3180627 1.3189720 1.2376667 1.1089961 
#>         9        10        11        12 
#> 0.9947792 1.0131536 1.0972773 1.0756162 
#> 
sd(x = ElNino_ERSST_region_1and2, method = "mode")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.6064859 0.5463920 0.6481940 0.8343168 0.8884459 0.8194681 0.6954894 0.6641566 
#>         9        10        11        12 
#> 0.6392201 0.6856086 0.7431749 0.7220048 
#> 
sd(x = ElNino_ERSST_region_1and2, method = "RP")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.6356319 0.5644116 0.6965143 0.9255739 0.9761812 0.9483744 0.8021007 0.7595808 
#>         9        10        11        12 
#> 0.6584798 0.6610759 0.6554421 0.6415014 
#> 
sd(x = ElNino_ERSST_region_1and2, method = "RPD")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.6212515 0.5734213 0.6938452 0.8249173 0.8652867 0.8293409 0.7414134 0.7098223 
#>         9        10        11        12 
#> 0.6579163 0.6803219 0.7254058 0.7231369 
#> 
sd(x = ElNino_ERSST_region_1and2, method = "radius", 
  alpha = 0.5, weight = "hard")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.6096900 0.5400084 0.6194035 0.7987683 0.8328768 0.7664151 0.6431316 0.6120882 
#>         9        10        11        12 
#> 0.6287866 0.6994844 0.7632923 0.7449797 
#> 
sd(x = ElNino_ERSST_region_1and2, method = "radius", 
  alpha = 0.5, weight = "soft")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.6090855 0.4880287 0.4821902 0.5698758 0.5284720 0.5239371 0.4533245 0.4789555 
#>         9        10        11        12 
#> 0.4880702 0.5085474 0.5808412 0.5696013 
#>