Simulated multiple sets of functional time series

We generate 2 groups of \(m\) functional time series. For each \(i\) in {1, ..., m} in a given cluster \(c\), \(c\) in {1,2}, the \(t\) th function, \(t\) in {1,..., T}, is given by $$Y_{it}^{(c)} (x)= \mu^{(c)}(x) + \sum_{k=1}^{2}\xi_{tk}^{(c)} \rho_k^{(c)} (x) + \sum_{l=1}^{2}\zeta_{itl}^{(c)} \psi_l^{(c)} (x) + \upsilon_{it}^{(c)} (x)$$

data("sim_ex_cluster")

Details

The mean functions for each of these two clusters are set to be \(\mu^{(1)}(x) = 2(x-0.25)^{2}\) and \(\mu^{(2)}(x) = 2(x-0.4)^{2}+0.1\).

While the variates \(\mathbf{\xi_{tk}^{(c)}}=(\xi_{1k}^{(c)}, \xi_{2k}^{(c)}, \ldots, \xi_{Tk}^{(c)})^{\top}\) for both clusters, are generated from autoregressive of order 1 with parameter 0.7, while the variates \(\zeta_{it1}^{(c)}\) and \(\zeta_{it2}^{(c)}\) for both clusters, are generated from independent and identically distributed \(N(0,0.5)\) and \(N(0,0.25)\), respectively.

The basis functions for the common-time trend for the first cluster, \(\rho_k^{(1)} (x)\), for \(k\) in {1,2} are \(sqrt(2)*sin(\pi*(0:200/200))\) and \(sqrt(2)*cos(\pi*(0:200/200))\) respectively; and the basis functions for the common-time trend for the second cluster, \(\rho_k^{(2)} (x)\), for \(k\) in {1,2} are \(sqrt(2)*sin(2\pi*(0:200/200))\) and \(sqrt(2)*cos(2\pi*(0:200/200))\) respectively.

The basis functions for the residual for the first cluster, \(\psi_l^{(1)} (x)\), for \(l\) in {1,2} are \(sqrt(2)*sin(3\pi*(0:200/200))\) and \(sqrt(2)*cos(3\pi*(0:200/200))\) respectively; and the basis functions for the residual for the second cluster, \(\psi_l^{(2)} (x)\), for \(l\) in {1,2} are \(sqrt(2)*sin(4\pi*(0:200/200))\) and \(sqrt(2)*cos(4\pi*(0:200/200))\) respectively.

The measurement error \(\upsilon_{it}\) for each continuum x is generated from independent and identically distributed \(N(0, 0.2^2)\)

Examples

data(sim_ex_cluster)