Detection of the optimal stopping time in a curve time series

Detecting the optimal stopping time for the glue curing of wood panels in an automatic process environment.

stop_time_detect(data, forecasting_method = c("ets", "arima", "rw"))

Arguments

data

An object of class fts

forecasting_method

A univariate time series forecasting method

Value

break_points_strucchange

Breakpoints detected by the regression approach

break_points_ecp

Breakpoints detected by the distance-based approach

err_forward

Forward integrated squared forecast errors

err_backward

Backward integrated squared forecast errors (ISFEs)

ncomp_select_forward

Number of components selected by the eigenvalue ratio tests based on the forward ISFEs

ncomp_select_backward

Number of components selected by the eigenvalue ratio tests based on the backward ISFEs

References

Bekhta, P., Ortynska, G. and Sedliacik, J. (2014). Properties of modified phenol-formaldehyde adhesive for plywood panels manufactured from high moisture content veneer. Drvna Industrija 65(4), 293-301.

Author

Han Lin Shang