Variance functions for functional time series

Computes variance functions of functional time series at each variable.

# S3 method for class 'fts'
var(x, method = c("coordinate", "FM", "mode", "RP", "RPD", "radius"), 
 trim = 0.25, alpha, weight, ...)

Arguments

x

An object of class fts.

method

Method for computing median.

trim

Percentage of trimming.

alpha

Tuning parameter when method="radius".

weight

Hard thresholding or soft thresholding.

...

Other arguments.

Details

If method = "coordinate", it computes coordinate-wise variance.

If method = "FM", it computes the variance of trimmed functional data ordered by the functional depth of Fraiman and Muniz (2001).

If method = "mode", it computes the variance of trimmed functional data ordered by \(h\)-modal functional depth.

If method = "RP", it computes the variance of trimmed functional data ordered by random projection depth.

If method = "RPD", it computes the variance of trimmed functional data ordered by random projection derivative depth.

If method = "radius", it computes the standard deviation function of trimmed functional data ordered by the notion of alpha-radius.

Value

A list containing x = variables and y = variance rates.

References

O. Hossjer and C. Croux (1995) "Generalized univariate signed rank statistics for testing and estimating a multivariate location parameter", Nonparametric Statistics, 4(3), 293-308.

A. Cuevas and M. Febrero and R. Fraiman (2006) "On the use of bootstrap for estimating functions with functional data", Computational Statistics and Data Analysis, 51(2), 1063-1074.

A. Cuevas and M. Febrero and R. Fraiman (2007), "Robust estimation and classification for functional data via projection-based depth notions", Computational Statistics, 22(3), 481-496.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2007) "A functional analysis of NOx levels: location and scale estimation and outlier detection", Computational Statistics, 22(3), 411-427.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2008) "Outlier detection in functional data by depth measures, with application to identify abnormal NOx levels", Environmetrics, 19(4), 331-345.

M. Febrero and P. Galeano and W. Gonzalez-Manteiga (2010) "Measures of influence for the functional linear model with scalar response", Journal of Multivariate Analysis, 101(2), 327-339.

J. A. Cuesta-Albertos and A. Nieto-Reyes (2010) "Functional classification and the random Tukey depth. Practical issues", Combining Soft Computing and Statistical Methods in Data Analysis, Advances in Intelligent and Soft Computing, 77, 123-130.

D. Gervini (2012) "Outlier detection and trimmed estimation in general functional spaces", Statistica Sinica, 22(4), 1639-1660.

Author

Han Lin Shang

See also

Examples

# Fraiman-Muniz depth was arguably the oldest functional depth.  
var(x = ElNino_ERSST_region_1and2, method = "FM")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.2822754 0.2884389 0.5057134 0.8767595 0.9192481 0.8400903 0.6371524 0.5738694 
#>         9        10        11        12 
#> 0.4815904 0.5275630 0.6741974 0.6753451 
#> 
var(x = ElNino_ERSST_region_1and2, method = "coordinate")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.8312815 0.6104055 0.8431023 1.3386591 1.7628376 1.7652707 1.5543457 1.2479587 
#>         9        10        11        12 
#> 1.0041384 1.0415756 1.2217236 1.1739643 
#> 
var(x = ElNino_ERSST_region_1and2, method = "mode")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.3678251 0.2985442 0.4201554 0.6960845 0.7893362 0.6715279 0.4837054 0.4411040 
#>         9        10        11        12 
#> 0.4086024 0.4700592 0.5523090 0.5212910 
#> 
var(x = ElNino_ERSST_region_1and2, method = "RP")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.3941630 0.3660735 0.5676859 0.8866809 0.9403290 0.8093572 0.5416149 0.5263173 
#>         9        10        11        12 
#> 0.4467726 0.4470193 0.4581399 0.4669827 
#> 
var(x = ElNino_ERSST_region_1and2, method = "RPD")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.3669519 0.3069564 0.4055774 0.6707980 0.7038515 0.6179566 0.4808398 0.4703351 
#>         9        10        11        12 
#> 0.4472664 0.4943282 0.5625898 0.5266448 
#> 
var(x = ElNino_ERSST_region_1and2, method = "radius", 
alpha = 0.5, weight = "hard")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.3717219 0.2916091 0.3836607 0.6380308 0.6936838 0.5873922 0.4136183 0.3746520 
#>         9        10        11        12 
#> 0.3953726 0.4892784 0.5826151 0.5549947 
#> 
var(x = ElNino_ERSST_region_1and2, method = "radius", 
alpha = 0.5, weight = "soft")
#> $x
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12
#> 
#> $y
#>         1         2         3         4         5         6         7         8 
#> 0.3709851 0.2381721 0.2325074 0.3247585 0.2792827 0.2745101 0.2055031 0.2293984 
#>         9        10        11        12 
#> 0.2382125 0.2586205 0.3373766 0.3244457 
#>